chore: benchmark ships the instrument only, internal tuning scripts removed
